Starting WriteStatsSQL class, bugs in Generate
Starting to write tests for WriteStatsSQL in 'spec/write_stats_sql_spec.rb'.
Decided to verify that the SQL from GenerateStatsSQL actually works and
logged into repository. Started irb, created a GenerateStatsSQL object and
started generating stats SQL strings. But it turns out I was dumb along the
way. Making multiple calls to the same object for the same or different SQL,
'jobs_started_this_year' or 'jobs_started_this_quarter', resulted in each
methods SQL being appended to the string -- so we end up with a string with
a bunch of extra 'where' clauses. Wrote tests to verify and make the issue
visible. Fixed by changing string concat operator '<<' to '+' in the methods.