Fleshing out basic WriteStatsSQL
Adding test to verify that each of the desired SQL statements is written
to the file as a newline-delimited list. Adding test to verify that each SQL
statement is only written to the file once (instead of constantly appending
new lines). Tests passing. Very much need to break out the list of SQL
statements to a file that can be shared amongst spec files.

Adding general outline of the algorithm to gather stats to 'notes.md'.