Adding functionality to run SQL via MySQL CLI
Since access to Repository is locked down to specific static IPs, it's
necessary to connect directly from Proc01. Unfortunately, Proc01 runs
Windows and I can't get the MySQL adapter for Ruby to compile correctly.
Instead we're going to hackily create a series of text files and pipe them
in and out of the MySQL command line client ( `mysql` ). Now that the SQL
commands are written to a file, that file will be piped to `mysql` and run
on Repository. The output will be saved for additional processing.

- notes.md

Checked off the first two steps and added a row of sample output

- write_stats_sql_spec.rb

Whitespace

- run_stats_sql_spec.rb

Initial tests for running `mysql` from the command line:
  Read the config from YAML
  Run the `mysql` client
  Capture the output of the `mysql` client