Argh. Everything is changing ...
In order to test the SQL infile functionality we need sample output to compare
the output to. However, the sample output will go out of date quickly unless we
can specify a certain (older, unchanging) file. So we need to start adding some
functionality for dealing with multiple files - a new one each day.